- Modify Error moved to a separate tsx file
