+++ spa/src/common/Wrapper.tsx
@@ -2,7 +2,7 @@ import styled from "@emotion/styled";
 
 export const Wrapper = styled.div`
 	margin: 0 auto;
-	max-width: 640px;
+	max-width: 580px;
 	height: calc(100vh - 56px);
 	display: flex;
 	flex-direction: column;
